+++ tests/unit/test_fisher_stats.py
@@ -1,5 +1,6 @@
 import cobra
 import numpy as np
+
 from como.stats.fisher_exact_test import FisherExactTest
 
 
@@ -10,8 +11,8 @@ def test_fisher_stats():
 
     assert real == FisherExactTest(
         pathway="Glycolysis/gluconeogenesis",
-        statistic=np.float64(4.321708185053381),
-        pvalue=np.float64(1.2883495211648955e-05),
+        statistic=4.321708185053381,
+        pvalue=1.2883495211648955e-05,
         a=32,
         b=10,
         c=4496,
